[TEST] make alphabet_tuple_base test a test template
