+++ test/unit/alphabet/structure/structured_aa_test.cpp
@@ -16,10 +16,48 @@
 
 #include "../alphabet_test_template.hpp"
 #include "../alphabet_constexpr_test_template.hpp"
+#include "../composite/alphabet_tuple_base_test_template.hpp"
 
 using namespace seqan3;
+
+template <>
+class alphabet_tuple_base_test<structured_aa<aa27, dssp9>>: public ::testing::Test
+{
+public:
+    using T = structured_aa<aa27, dssp9>;
+
+    T instance = T{value_1(), value_2()};
+    T zero_instance = T{decltype(value_1()){}, decltype(value_2()){}};
+    size_t tup_size{2};
+
+    // structured_aa<aa27, dssp9>
+    // -------------------------------------------------------------------------
+    aa27 value_1()
+    {
+        return 'K'_aa27;
+    }
+    aa27 assignable_to_value_1()
+    {
+        return 'K'_aa27; // replace if assignable subtype becomes available
+    }
+    dssp9 value_2()
+    {
+        return 'I'_dssp9;
+    }
+    dssp9 assignable_to_value_2()
+    {
+        return 'I'_dssp9; // replace if assignable subtype becomes available
+    }
+    auto values_to_cmp()
+    {
+        return std::make_tuple(/*low */'A'_aa27, 'H'_dssp9,
+                               /*mid */'P'_aa27, 'I'_dssp9,
+                               /*high*/'Z'_aa27, 'X'_dssp9);
+    }
+};
 
 using structured_aa_types = ::testing::Types<structured_aa<aa27, dssp9>>;
 
 INSTANTIATE_TYPED_TEST_CASE_P(structured_aa, alphabet, structured_aa_types);
 INSTANTIATE_TYPED_TEST_CASE_P(structured_aa, alphabet_constexpr, structured_aa_types);
+INSTANTIATE_TYPED_TEST_CASE_P(structured_aa, alphabet_tuple_base_test, structured_aa_types);
